+++ docs/src/conf.py
@@ -23,7 +23,7 @@ sys.path.insert(0, os.path.abspath('../../'))
 project = 'thelper'
 copyright = '2018, Pierre-Luc St-Charles'
 author = 'Pierre-Luc St-Charles'
-version = release = '0.6.1'
+version = release = '0.6.2'
 
 
 # -- General configuration ---------------------------------------------------