+++ thelper/__init__.py
@@ -23,4 +23,4 @@ import thelper.viz  # noqa: F401
 logger = logging.getLogger("thelper")
 
 __url__ = "https://github.com/plstcharles/thelper"
-__version__ = "0.6.1"
+__version__ = "0.6.2"