+++ CHANGELOG.rst
@@ -6,6 +6,11 @@ Changelog
 `Unreleased <http://github.com/plstcharles/thelper/tree/master>`_ (latest)
 ----------------------------------------------------------------------------------
 
+.. **INSERT APPLIED CHANGES HERE**
+
+`0.6.2 <http://github.com/plstcharles/thelper/tree/v0.6.2>`_ (2020/10/01)
+----------------------------------------------------------------------------------
+
 * Added ``SelectChannels`` transform operator for inplace sample modifications that might need it.
 
 `0.6.1 <http://github.com/plstcharles/thelper/tree/v0.6.1>`_ (2020/07/29)